[TH2-5004] refactored after review
