+++ cradle-core/src/main/java/com/exactpro/cradle/CoreStorageSettings.java
@@ -16,10 +16,21 @@
 
 package com.exactpro.cradle;
 
+import com.exactpro.cradle.messages.GroupedMessageBatchToStore;
+import com.exactpro.cradle.messages.MessageBatchToStore;
+
 @SuppressWarnings("unused")
 public class CoreStorageSettings {
     public static final long DEFAULT_BOOK_REFRESH_INTERVAL_MILLIS = 60000;
-    protected long bookRefreshIntervalMillis = DEFAULT_BOOK_REFRESH_INTERVAL_MILLIS;
+    public static final boolean DEFAULT_STORE_INDIVIDUAL_MESSAGE_SESSIONS = true;
+    private long bookRefreshIntervalMillis = DEFAULT_BOOK_REFRESH_INTERVAL_MILLIS;
+
+    /**
+     * if `true`, the {@link CradleStorage#storeGroupedMessageBatch} method converts {@link GroupedMessageBatchToStore} to {@link MessageBatchToStore} and stores them.
+     * if `false`, the {@link CradleStorage#storeGroupedMessageBatch} method stores only {@link GroupedMessageBatchToStore},
+     * also the {@link CradleStorage#storeMessageBatch} and {@link CradleStorage#storeMessageBatchAsync} methods throws {@link  IllegalStateException}
+     */
+    private boolean storeIndividualMessageSessions = DEFAULT_STORE_INDIVIDUAL_MESSAGE_SESSIONS;
 
     public long getBookRefreshIntervalMillis() {
         return bookRefreshIntervalMillis;
@@ -35,5 +46,13 @@ public class CoreStorageSettings {
 
     public long calculateStoreActionRejectionThreshold() {
         return this.bookRefreshIntervalMillis;
+    }
+
+    public boolean isStoreIndividualMessageSessions() {
+        return storeIndividualMessageSessions;
+    }
+
+    public void setStoreIndividualMessageSessions(boolean storeIndividualMessageSessions) {
+        this.storeIndividualMessageSessions = storeIndividualMessageSessions;
     }
 }
